requirements.txt
requests>=2.28.0
Pillow>=9.0.0
numpy>=1.23.0
torch>=2.0.0
